fix the conflicts(#130)
